refactor(add-rework): move exercise card widgets to exercise_card directory
re #105
